+++ src/braket/experimental/autoqasm/types/__init__.py
@@ -15,7 +15,7 @@
 for type handling.
 """
 
-from .py_to_oqpy import map_type, wrap_value  # noqa: F401
+from .conversions import map_type, wrap_value  # noqa: F401
 from .types import (  # noqa: F401
     ArrayVar,
     BitVar,
